backend/requirements.txt
Flask
python-dotenv
google-generativeai
requests
Flask-Cors